Asesorias/requirements.txt
Django 
Djangorestframework
Djangorestframework-simplejwt
Django-model-utils
django-jsonfield
psycopg2-binary
gunicorn
